requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
wtforms==3.1.1
mysql-connector-python==8.1.0
Werkzeug==3.0.1
